Removed dummy.txt from the home folder as it had fulfilled its purpose.
Added a logs directory to the home folder to record each day's work, starting
from the day I acquired ownership of this repository.

The logs will all be html files, and will be named with the date on which they
were created. A log file created on the 2nd of August in 2018 will have the
name august_2.html. Also, all logs will use the stylesheet log_style.css
which is stored in the stylesheets subdirectory of logs.

I added the log for the 2nd of August 2018 and also added the stylesheet
log_style.css.